+++ test/server/src/parser.test.ts
@@ -1332,6 +1332,22 @@ describe("Parser", () => {
 			expect(received[0].location.range.end.line).to.equal(21);
 		});
 
+		it("should callback on a reference command as part of a choice option", () => {
+			let fakeDocument = createDocument("*choice\n\t*hide_reuse #Option ${variable}");
+			let received: Array<Symbol> = [];
+			let fakeCallbacks = Substitute.for<ParserCallbacks>();
+			fakeCallbacks.onVariableReference(Arg.all()).mimicks((s: string, l: Location, state: ParsingState) => {
+				received.push({text: s, location: l});
+			});
+	
+			parse(fakeDocument, fakeCallbacks);
+	
+			expect(received.length).to.equal(1);
+			expect(received[0].text).to.equal("variable");
+			expect(received[0].location.range.start.line).to.equal(31);
+			expect(received[0].location.range.end.line).to.equal(39);
+		});
+
 		it("should callback on a variable in an *if before an option", () => {
 			let fakeDocument = createDocument("*choice\n\t*if variable # This is a choice");
 			let received: Array<Symbol> = [];